Added loadProjectsFile(), saveProjectsfile() bodies
In the near future those methods are going to be used to provide
a proper communication between projects.dat file and application
(it'll use new FileManager class and Project serialization methods).

Before we replace old methods with those, we have to change some
logic that was used by the old ones (i.e. reading whole projects.dat
file and splitting it by newline chars) and make it more clever.

Change-Id: Ib275497ec6024077e3b44113221b1a8e52a8c4b7